renamed and documented DashboardState
