+++ lib/src/layouts/dashboard/dashboard.dart
@@ -1,4 +1,4 @@
-export 'bloc/admin_panel_bloc.dart';
+export 'bloc/dashboard_bloc.dart';
 export 'repositories/repositories.dart';
 export 'views/views.dart';
 export 'widgets/widgets.dart';
