+++ lib/src/layouts/dashboard/widgets/side_menu_section.dart
@@ -66,7 +66,7 @@ class SideMenuSection extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<AdminPanelBloc, AdminPanelState>(
+    return BlocBuilder<AdminPanelBloc, DashboardState>(
       builder: (context, state) {
         return Column(
           mainAxisSize: MainAxisSize.min,
