+++ lib/src/layouts/dashboard/widgets/side_menu_sub_item.dart
@@ -57,7 +57,7 @@ class SideMenuSubItem extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<AdminPanelBloc, AdminPanelState>(
+    return BlocBuilder<AdminPanelBloc, DashboardState>(
       builder: (context, state) {
         bool isHoverOrActive() {
           return id! == state.sideMenuItemBeingHovered ||
