+++ test/src/flutter_admin_panel_test.dart
@@ -1,6 +1,5 @@
 // ignore_for_file: prefer_const_constructors
 
-import 'package:flexilayout/flexilayout.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
